requirements.txt
nemo_toolkit['all']
pyannote.audio
pydub
